move test utils tests in their own new module
